requirements.txt
discord.py>=2.3.2
aiosqlite>=0.17.0
python-dotenv>=0.19.0
mcstatus>=11.0.0
mysql-connector-python>=8.0.0
PyYAML>=6.0